feat(run code with exmaple test cases in DB)
added methods inside codeRunService. is problem is 0 we are considering it as palin code and will run that with user sent input else we query for test cases related to that problem id and set input as that tc's to run.
currenlty not cheking if the problem id is valid or not and excepting validation code inside request body.